refactor: fixes #3
this separates the controller and handler functions into separate
files with respective functions. aims to make the codebase clean.
